Remove jquery-ui-themes gem, but leave the assets
As I only need the theme for the custom datetimepicker control, I don't
see the value in keeping this gem around.
It would be worth keeping if the whole site was based on the theme,
because then I could imagine changing it quite often, whereas now I
probably won't need to change it very often and the deprication warnings
this gem produces are annoying!